requirements.txt
python-dotenv>=1.0.0
langchain>=0.2.0
langchain_community>=0.2.0
langchain_groq>=0.1.0
langgraph>=0.2.0
chromadb>=0.5.0
streamlit>=1.30.0
PyPDF2>=3.0.1